New Device Type: Star Projector
Added new device type: Star Projector

This adds support for the BlissLights Sky Lite Evolve Star Projector

https://www.amazon.com/dp/B0B4VPM54L?th=1

Essentially an RGB light with a switch for the rotation motor. The rotation could be broken out into a brightness slider, but at max speed the movement is too subtle for me to justify the effort. Other models come with laser stars, which would require additional switches.

DO NOT UPDATE FIRMWARE!!
I'm unsure if the update moved to a Tuya 3.4+ scheme, or just changed something in the packet format, but it became unusable after updating.